requirements.txt
absl-py==0.12.0
apache-beam==2.30.0
astunparse==1.6.3
attrs==21.2.0
avro-python3==1.9.2.1
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
contextlib2==0.6.0.post1
crcmod==1.7
cycler==0.10.0
Cython==0.29.23
dill==0.3.1.1
dm-tree==0.1.6
docopt==0.6.2
fastavro==1.4.1
flatbuffers==1.12
future==0.18.2
gast==0.4.0
gin-config==0.4.0
google-api-core==1.30.0
google-api-python-client==2.9.0
google-auth==1.31.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-cloud-bigquery==2.20.0
google-cloud-core==1.7.0
google-crc32c==1.1.2
google-pasta==0.2.0
google-resumable-media==1.3.1
googleapis-common-protos==1.53.0
grpcio==1.34.1
h5py==3.1.0
hdfs==2.6.0
html5lib==1.1
httplib2==0.19.1
idna==2.10
importlib-resources==5.1.4
joblib==1.0.1
kaggle==1.5.12
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lvis==0.5.3
lxml==4.6.3
Markdown==3.3.4
matplotlib==3.4.2
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.1.1
opencv-python==4.5.2.54
opencv-python-headless==4.5.2.54
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.4
Pillow==8.2.0
portalocker==2.0.0
promise==2.3
proto-plus==1.18.1
protobuf==3.17.3
psutil==5.8.0
py-cpuinfo==8.0.0
pyarrow==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycparser==2.20
pydot==1.4.2
pymongo==3.11.4
pyparsing==2.4.7
pyserial==3.5
python-dateutil==2.8.1
python-slugify==5.0.2
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
sacrebleu==1.5.1
scikit-learn==0.24.2
scipy==1.7.0
sentencepiece==0.1.96
seqeval==1.2.2
six==1.15.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-addons==0.13.0
tensorflow-datasets==4.3.0
tensorflow-estimator==2.5.0
tensorflow-hub==0.12.0
tensorflow-metadata==1.0.0
tensorflow-model-optimization==0.6.0
termcolor==1.1.0
text-unidecode==1.3
tf-models-official==2.5.0
tf-slim==1.1.0
threadpoolctl==2.1.0
tqdm==4.61.1
typeguard==2.12.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.5
webencodings==0.5.1
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.4.1